requirements.txt
altair==4.2.2
annotated-types==0.7.0
appdirs==1.4.4
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
black==25.1.0
boto3==1.40.11
botocore==1.40.11
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
coverage==7.10.3
cramjam==2.11.0
cryptography==45.0.6
distro==1.9.0
entrypoints==0.4
exceptiongroup==1.3.0
fastparquet==2024.11.0
flake8==7.3.0
fsspec==2025.7.0
great_expectations==1.5.8
h11==0.16.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
marshmallow==3.26.1
mccabe==0.7.0
mistune==3.1.3
moto==5.1.10
mypy_extensions==1.1.0
nfl_data_py==0.3.3
numpy==1.26.4
outcome==1.3.0.post0
packaging==25.0
pandas==1.5.3
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
posthog==6.5.0
py4j==0.10.9.9
pyarrow==21.0.0
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.2.3
PySocks==1.7.1
pyspark==4.0.0
pytest==8.4.1
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
responses==0.25.8
rpds-py==0.27.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
s3transfer==0.13.1
scipy==1.13.1
selenium==4.32.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
structlog==25.4.0
tomli==2.2.1
toolz==1.0.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzlocal==5.3.1
urllib3==1.26.20
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
xmltodict==0.14.2
